Ensure correct timestamp type in Message
Fixes #118

When an incorrect type is provided by the user for Message.timestamp or
Message.next_timestamp, MsgPack will serialize an invalid MMPMessage on
the wire. This leads to errors in statically typed peer actors.

Issue is fixed by explicitly converting to float in Message.__init__ and
checking again in MMPMessage.__init__ (as the user may have assigned
another value between creation of the Message and Instance.send).